+++ app/src/main/java/calculator/exception/ExpressionInputException.java
@@ -1,4 +1,4 @@
-package calculator.validation.exception;
+package calculator.exception;
 
 public class ExpressionInputException extends RuntimeException {
